Make edit to today actions bottom sheet

+++ lib/features/everyday/presentation/components/today_actions_bottom_sheet.dart
@@ -1,16 +1,32 @@
 import 'dart:io';
 
 import 'package:flutter/material.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:gap/gap.dart';
 import 'package:myapp/core/extensions.dart';
+import 'package:myapp/core/resources/data_state.dart';
 import 'package:myapp/features/everyday/domain/entities/today.dart';
+import 'package:myapp/features/everyday/presentation/components/today_caption_dialog.dart';
+import 'package:myapp/features/everyday/presentation/providers/today_provider.dart';
 
-class TodayActionsBottomSheet extends StatelessWidget {
+class TodayActionsBottomSheet extends ConsumerStatefulWidget {
   const TodayActionsBottomSheet({
     super.key,
     required this.today,
   });
   final Today today;
+
+  @override
+  ConsumerState<TodayActionsBottomSheet> createState() =>
+      _TodayActionsBottomSheetState();
+}
+
+class _TodayActionsBottomSheetState
+    extends ConsumerState<TodayActionsBottomSheet> {
+  late String _displayCaption = widget.today.caption;
+  late DateTime _displayDateTime = widget.today.date;
+
+  bool _isUpdating = false;
   @override
   Widget build(BuildContext context) {
     return Column(
@@ -29,19 +45,21 @@ class TodayActionsBottomSheet extends StatelessWidget {
               ),
               TodayAction(
                 icon: Icons.delete_outline,
-                text: 'Delete',
+                text: 'Permanent Delete',
                 onTap: () {},
               ),
-              TodayAction(
-                icon: Icons.backup_outlined,
-                text: 'Back up',
-                onTap: () {},
-              ),
-              TodayAction(
-                icon: Icons.mobile_off_outlined,
-                text: 'Delete from device',
-                onTap: () {},
-              ),
+              if (!widget.today.isBackedUp)
+                TodayAction(
+                  icon: Icons.backup_outlined,
+                  text: 'Back up',
+                  onTap: () {},
+                ),
+              if (widget.today.isAvailableLocal)
+                TodayAction(
+                  icon: Icons.mobile_off_outlined,
+                  text: 'Delete from device',
+                  onTap: () {},
+                ),
             ],
           ),
         ),
@@ -59,38 +77,75 @@ class TodayActionsBottomSheet extends StatelessWidget {
                 children: [
                   ListTile(
                     leading: const Icon(Icons.text_snippet),
-                    title: Text(today.caption),
+                    title: Text(_displayCaption),
                     trailing: IconButton(
                       icon: const Icon(Icons.edit),
-                      onPressed: () {
-                        // showDialog(
-                        //     context: context,
-                        //     builder: (context) =>
-                        //         const TodayCaptionDialog(videoPath: ''));
+                      onPressed: () async {
+                        final result = await showDialog(
+                          context: context,
+                          builder: (context) =>
+                              TodayCaptionDialog(today: widget.today),
+                        );
+                        if (result != null) {
+                          setState(() {
+                            _displayCaption = result;
+                          });
+                        }
                       },
                     ),
                   ),
                   ListTile(
                     leading: const Icon(Icons.calendar_month),
-                    title: Text(today.date.formatDateWithShortDayWithTime),
+                    title:
+                        Text(_displayDateTime.formatDateWithShortDayWithTime),
                     trailing: IconButton(
-                      icon: const Icon(Icons.edit),
+                      icon: _isUpdating
+                          ? const SizedBox.square(
+                              dimension: 24,
+                              child: CircularProgressIndicator(),
+                            )
+                          : const Icon(Icons.edit),
                       onPressed: () async {
                         final date = await showDatePicker(
                           context: context,
-                          currentDate: today.date,
-                          firstDate:
-                              today.date.subtract(const Duration(days: 14)),
+                          currentDate: widget.today.date,
+                          firstDate: widget.today.date
+                              .subtract(const Duration(days: 14)),
                           lastDate: DateTime.now(),
                         );
                         if (date != null) {
                           if (context.mounted) {
-                            showTimePicker(
+                            final time = await showTimePicker(
                               context: context,
                               initialTime: TimeOfDay(
-                                  hour: today.date.hour,
-                                  minute: today.date.minute),
+                                  hour: widget.today.date.hour,
+                                  minute: widget.today.date.minute),
                             );
+                            setState(() {
+                              _isUpdating = true;
+                            });
+
+                            if (time != null) {
+                              final dateTime = DateTime(
+                                date.year,
+                                date.month,
+                                date.day,
+                                time.hour,
+                                time.minute,
+                              );
+                              final dataState = await ref
+                                  .read(todayProvider.notifier)
+                                  .updateToday(
+                                      widget.today.copyWith(date: dateTime));
+                              if (dataState is DataSuccess) {
+                                setState(() {
+                                  _displayDateTime = dateTime;
+                                });
+                              }
+                              setState(() {
+                                _isUpdating = false;
+                              });
+                            }
                           }
                         }
                       },
@@ -98,10 +153,10 @@ class TodayActionsBottomSheet extends StatelessWidget {
                   ),
                   ListTile(
                     leading: const Icon(Icons.mobile_friendly),
-                    title: !today.isAvailableLocal
+                    title: !widget.today.isAvailableLocal
                         ? const Text('Not available offline')
                         : FutureBuilder(
-                            future: File(today.localVideoPath!).size(1),
+                            future: File(widget.today.localVideoPath!).size(1),
                             builder: (context, snapshot) {
                               if (snapshot.connectionState ==
                                   ConnectionState.waiting) {
@@ -114,11 +169,12 @@ class TodayActionsBottomSheet extends StatelessWidget {
                   ),
                   ListTile(
                     leading: const Icon(Icons.backup_outlined),
-                    title: !today.isBackedUp
+                    title: !widget.today.isBackedUp
                         ? const Text('Not Backed up')
-                        : today.isAvailableLocal
+                        : widget.today.isAvailableLocal
                             ? FutureBuilder(
-                                future: File(today.localVideoPath!).size(1),
+                                future:
+                                    File(widget.today.localVideoPath!).size(1),
                                 builder: (context, snapshot) {
                                   if (snapshot.connectionState ==
                                       ConnectionState.waiting) {
@@ -129,17 +185,6 @@ class TodayActionsBottomSheet extends StatelessWidget {
                                 },
                               )
                             : const Text('Backed up'),
-                  ),
-                  ListTile(
-                    leading: const Icon(Icons.location_on_outlined),
-                    title: Row(
-                      children: [
-                        TextButton(
-                          onPressed: () {},
-                          child: const Text('Add a location'),
-                        ),
-                      ],
-                    ),
                   ),
                 ],
               )
